Suggest adding a build.properties file.
When I tried to build socko, starting sbt failed because the sbt runner
script (https://raw.github.com/paulp/sbt-extras/master/sbt) detected the
project as needing sbt 0.11.0.  The project depends on version 0.6 of
the xsbt gpg plugin, which is only available for 0.11.2.  Adding a
project/build.properties allows the runner script to know exactly which
version of sbt is required.